feat!: Implement character search and history display
This commit adds the core functionality for searching characters using the Raider.IO API and displaying their history.  The API call is functioning correctly and returns a JSON response which is stored locally.  However, an error is currently occurring after the successful API call which prevents the data from being correctly displayed to the user.  Further development will be needed to handle this exception.